Added a way to just install cursor from the repos if it already exists.

+++ core/tabs/applications-setup/developer-tools/cursor.sh
@@ -7,22 +7,33 @@ installCursor() {
         printf "%b\n" "${YELLOW}Installing Cursor...${RC}"
         case $PACKAGER in
             apt-get|nala)
-                TEMP_DEB="$(wd)"
-                curl -sSLo "$TEMP_DEB" 'https://api2.cursor.sh/updates/download/golden/linux-x64-deb/cursor/latest'
+                # Sometimes in ubuntu, the cursor package already exists.
+                if apt-cache search "^${PACKAGE_NAME}$" | grep -q "^${PACKAGE_NAME} -"; then
+                    TEMP_DEB="cursor.deb"
 
-                "$ESCALATION_TOOL" "$PACKAGER" update
-                "$ESCALATION_TOOL" "$PACKAGER" install -y "$TEMP_DEB"
-                rm "$TEMP_DEB"
+                    curl -sSLo "$TEMP_DEB" 'https://api2.cursor.sh/updates/download/golden/linux-x64-deb/cursor/latest'
+
+                    "$ESCALATION_TOOL" "$PACKAGER" update
+                    "$ESCALATION_TOOL" "$PACKAGER" install -y "$TEMP_DEB"
+                    rm "$TEMP_DEB"
+                else
+                    "$ESCALATION_TOOL" "$PACKAGER" install -y cursor
+                fi
+
                 ;;
             pacman)
                 "$AUR_HELPER" -S --needed --noconfirm cursor-bin
                 ;;
             dnf)
-                TEMP_RPM="$(wd)"
-                curl -sSLo "$TEMP_RPM" "https://api2.cursor.sh/updates/download/golden/linux-x64-rpm/cursor/latest"
+                if dnf list available "cursor" > /dev/null; then
+                    "$ESCALATION_TOOL" "$PACKAGER" install -y cursor
+                else
+                    TEMP_RPM="cursor.rpm"
+                    curl -sSLo "$TEMP_RPM" "https://api2.cursor.sh/updates/download/golden/linux-x64-rpm/cursor/latest"
 
-                "$ESCALATION_TOOL" "$PACKAGER" install -y "$TEMP_RPM"
-                rm "$TEMP_RPM"
+                    "$ESCALATION_TOOL" "$PACKAGER" install -y "$TEMP_RPM"
+                    rm "$TEMP_RPM"
+                fi
                 ;;
             *)
                 printf "%b\n" "${RED}Unsupported package manager: ${PACKAGER}${RC}"
